extract nested class in DicomDump
